trying (and failing) to add to cart
Most of the work in this commit is in the searchcontroller.  specifically inside of the AddToCart method.  It's very messy, because I've been trying different things, and I don't like deleting, so I usually just comment things out. The AddToCart view is not done, it is just a renamed copy of the search view.  But My Linq queries have compiled, and I was hoping that they would work to add the product to the database.  Tomorrow I will look into trying some more tutorials, and trying to change the AddToCart view to see if it fixes anything.