Still need to be able to add stuff to ShoppingCartProductTable
Made UserName in login.  Some reason Only the username seems to be required.  If I try using only the email address, it throws an error, but if I try using only the username, it’s okay.  Could not figure out why. (maybe try making the method async?...)

Updated the register controller so that it adds the user to my database.
It now also automatically adds a shopping cart for each  new user.
I’m trying to use AddToCart3 method inside of SearchController, but is not working, and I have tried many things.  That is where I am at now.